Feat/add slr to frontend

+++ engine/engineapi/slr/predictions/individualPredictions.py
@@ -46,12 +46,12 @@ def customSlrIndividualPrediction(file, predictor, response, b0, b1, xToPredict)
         raise ValueError("Invalid X to predict")
     
     try:
-        b0int = int(b0)
+        b0int = float(b0)
     except Exception:
         raise ValueError("Invalid b0")
     
     try:
-        b1int = int(b1)
+        b1int = float(b1)
     except Exception:
         raise ValueError("Invalid b1")
